+++ server/src/core/file-tracker/index.ts
@@ -1,1 +1,2 @@
 export * from './file-tracker'
+export * from './tracking-map'