+++ server/src/hover.ts
@@ -32,7 +32,7 @@ export async function findHover(
 			return null
 		}
 
-		return await findHoverInAny(fromPart, currentHTMLService, document, cssServiceMap, configuration)
+		return await findHoverInHTML(fromPart, currentHTMLService, document, htmlServiceMap, cssServiceMap, configuration)
 	}
 	else if (isCSSFile) {
 		let currentCSSService = await cssServiceMap.forceGetServiceByDocument(document)
@@ -45,7 +45,7 @@ export async function findHover(
 			return null
 		}
 
-		return await findHoverInAny(fromPart, currentCSSService, document, cssServiceMap, configuration)
+		return await findHoverInCSS(fromPart, currentCSSService, document, cssServiceMap, configuration)
 	}
 
 	return null
@@ -53,7 +53,54 @@ export async function findHover(
 
 
 /** Find hover in HTML or CSS files. */
-async function findHoverInAny(
+async function findHoverInHTML(
+	fromPart: Part,
+	currentService: HTMLService | CSSService,
+	document: TextDocument,
+	htmlServiceMap: HTMLServiceMap,
+	cssServiceMap: CSSServiceMap,
+	configuration: Configuration
+): Promise<Hover | null> {
+	if (!fromPart.isReferenceType()) {
+		return null
+	}
+
+	let hover: Hover | null = null
+	let matchPart = PartConvertor.toDefinitionMode(fromPart)
+
+
+	if (fromPart.isSelectorType() || fromPart.isCSSVariableType()) {
+
+		// Find within current document.
+		hover = await findEmbeddedOrImported(matchPart, fromPart, currentService, document, cssServiceMap, configuration)
+		if (hover) {
+			return hover
+		}
+
+
+		// Find across all css documents.
+		hover = await cssServiceMap.findHover(matchPart, fromPart, document, configuration.maxHoverStylePropertyCount)
+		if (hover) {
+			return hover
+		}
+
+
+		// Find across all html documents.
+		if (configuration.enableSharedCSSFragments) {
+			hover = await htmlServiceMap.findHover(matchPart, fromPart, document, configuration.maxHoverStylePropertyCount)
+			if (hover) {
+				return hover
+			}
+		}
+	}
+
+	return null
+}
+
+
+
+/** Find hover in HTML or CSS files. */
+async function findHoverInCSS(
 	fromPart: Part,
 	currentService: HTMLService | CSSService,
 	document: TextDocument,
@@ -64,25 +111,60 @@ async function findHoverInAny(
 		return null
 	}
 
-
+	let hover: Hover | null = null
 	let matchPart = PartConvertor.toDefinitionMode(fromPart)
 
+	if (fromPart.isCSSVariableType()) {
 
-	// Find within current document.
+		// Find within current document.
+		hover = await findEmbeddedOrImported(matchPart, fromPart, currentService, document, cssServiceMap, configuration)
+		if (hover) {
+			return hover
+		}
+
+
+		// Find across all css documents.
+		hover = await cssServiceMap.findHover(matchPart, fromPart, document, configuration.maxHoverStylePropertyCount)
+		if (hover) {
+			return hover
+		}
+	}
+	return null
+}
+
+
+async function findEmbeddedOrImported(
+	matchPart: Part,
+	fromPart: Part,
+	currentService: HTMLService | CSSService,
+	document: TextDocument,
+	cssServiceMap: CSSServiceMap,
+	configuration: Configuration
+): Promise<Hover | null> {
+
+	// Find embedded hover.
 	let hover = currentService.findHover(matchPart, fromPart, document, configuration.maxHoverStylePropertyCount)
 	if (hover) {
 		return hover
 	}
+	
 
+	// Having CSS files imported, firstly search within these files, if found, not searching more.
+	let cssURIs = await currentService.getImportedCSSURIs()
+	let cssURIChain = cssServiceMap.trackingMap.resolveChainedImportedURIs(cssURIs)
 
-	// Find across all css documents.
-	if (fromPart.isSelectorType() || fromPart.isCSSVariableType()) {
-		hover = await cssServiceMap.findHover(matchPart, fromPart, document, configuration.maxHoverStylePropertyCount)
+	for (let cssURI of cssURIChain) {
+		let cssService = await cssServiceMap.forceGetServiceByURI(cssURI)
+		if (!cssService) {
+			continue
+		}
+
+		let hover = cssService.findHover(matchPart, fromPart, document, configuration.maxHoverStylePropertyCount)
+		if (hover) {
+			return hover
+		}
 	}
 
-	if (hover) {
-		return hover
-	}
 
 	return null
 }