+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # ChangeLog
 
 
+## [2.7.0]
+
+- Provides `enableSharedCSSFragments` option, if your HTML codes contain imported CSS or CSS fragments which should be shared across whole workspace, you can set it to `true`.
+- Provides css class name reference diagnostic service, can set `enableClassNameReferenceDiagnostic=true` to enable it.
+- Changes `enableClassNameDiagnostic` to `enableClassNameDefinitionDiagnostic`.
+
+
 ## [2.6.0]
 
 - For #107, supports class name diagnostic.